PARSE: remove redefinition check in parser
