+++ base/src/main/java/org/aya/concrete/resolve/error/RedefinitionPrimError.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.concrete.error;
+package org.aya.concrete.resolve.error;
 
 import org.aya.api.distill.DistillerOptions;
 import org.aya.api.error.Problem;
@@ -9,21 +9,20 @@ import org.aya.pretty.doc.Style;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
 
-public record UnknownPrimError(
-  @Override @NotNull SourcePos sourcePos,
-  @NotNull String name
+public record RedefinitionPrimError(
+  @NotNull String name,
+  @Override @NotNull SourcePos sourcePos
 ) implements Problem {
   @Override public @NotNull Doc describe(@NotNull DistillerOptions options) {
-    return Doc.sep(
-      Doc.plain("Unknown primitive"),
+    return Doc.sep(Doc.plain("Redefinition of primitive"),
       Doc.styled(Style.code(), Doc.plain(name)));
-  }
-
-  @Override public @NotNull Stage stage() {
-    return Stage.PARSE;
   }
 
   @Override public @NotNull Severity level() {
     return Severity.ERROR;
+  }
+
+  @Override public @NotNull Stage stage() {
+    return Stage.PARSE;
   }
 }
