+++ base/src/main/java/org/aya/concrete/resolve/visitor/StmtShallowResolver.java
@@ -6,9 +6,6 @@ import kala.collection.SeqLike;
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableHashMap;
 import kala.tuple.Tuple2;
-import org.aya.concrete.error.PrimDependencyError;
-import org.aya.concrete.error.RedefinitionError;
-import org.aya.concrete.error.UnknownPrimError;
 import org.aya.concrete.remark.Remark;
 import org.aya.concrete.resolve.ResolveInfo;
 import org.aya.concrete.resolve.context.Context;
@@ -16,6 +13,9 @@ import org.aya.concrete.resolve.context.ModuleContext;
 import org.aya.concrete.resolve.context.NoExportContext;
 import org.aya.concrete.resolve.context.PhysicalModuleContext;
 import org.aya.concrete.resolve.error.ModNotFoundError;
+import org.aya.concrete.resolve.error.PrimDependencyError;
+import org.aya.concrete.resolve.error.RedefinitionPrimError;
+import org.aya.concrete.resolve.error.UnknownPrimError;
 import org.aya.concrete.resolve.module.ModuleLoader;
 import org.aya.concrete.stmt.*;
 import org.aya.core.def.PrimDef;
@@ -126,7 +126,7 @@ public record StmtShallowResolver(
         if (lack.isNotEmpty() && lack.get().isNotEmpty())
           context.reportAndThrow(new PrimDependencyError(name, lack.get(), sourcePos));
         else if (PrimDef.Factory.INSTANCE.have(primID))
-          context.reportAndThrow(new RedefinitionError(RedefinitionError.Kind.Prim, name, sourcePos));
+          context.reportAndThrow(new RedefinitionPrimError(name, sourcePos));
         PrimDef.Factory.INSTANCE.factory(primID, decl.ref);
         resolveDecl(decl, context);
       }
