+++ base/src/test/resources/failure/scope/issue274.aya.txt
@@ -5,8 +5,8 @@ In file $FILE:3:4 ->
   3 |   | zero
           ^--^
 
-Error: Redefinition of constructor `zero`
+Error: The name zero (`zero`) is already defined elsewhere
 
-Parsing interrupted due to:
+Resolving interrupted due to:
 1 error(s), 0 warning(s).
 What are you doing?
